backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6
fastembed==0.4.1
groq==0.12.0
resend==2.4.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-httpx==0.35.0
